requirements.txt
python-dotenv
google-genai
openai